refactor compile time constants
