+++ tardis/opacities/opacities.py
@@ -6,7 +6,7 @@ from tardis import constants as const
 from tardis.transport.montecarlo import (
     njit_dict_no_parallel,
 )
-from tardis.transport.montecarlo.numba_config import (
+from tardis.transport.montecarlo.configuration.constants import (
     SIGMA_THOMSON,
 )
