+++ tardis/transport/montecarlo/estimators/radfield_estimator_calcs.py
@@ -5,7 +5,7 @@ from numba import njit
 from tardis.transport.montecarlo import (
     njit_dict_no_parallel,
 )
-from tardis.transport.montecarlo.numba_config import KB, H
+from tardis.transport.montecarlo.configuration.constants import KB, H
 from tardis.transport.frame_transformations import (
     calc_packet_energy,
     calc_packet_energy_full_relativity,
