--- tardis/transport/montecarlo/montecarlo_globals.py
@@ -1,9 +1,0 @@
-ENABLE_FULL_RELATIVITY = False
-LINE_INTERACTION_TYPE = 0
-DISABLE_ELECTRON_SCATTERING = False
-DISABLE_LINE_SCATTERING = False
-
-LEGACY_MODE_ENABLED = False
-ENABLE_RPACKET_TRACKING = False
-CONTINUUM_PROCESSES_ENABLED = False
-ENABLE_VPACKET_TRACKING = False
